+++ GraphicLegends/src/main/kotlin/Interfaces/IColorSpace.kt
@@ -1,11 +1,11 @@
 package Interfaces
 
 interface IColorSpace {
-    fun ToRGB(values : DoubleArray) : DoubleArray
-    fun ToCMY(values : DoubleArray) : DoubleArray
-    fun ToHSL(values : DoubleArray) : DoubleArray
-    fun ToHSV(values: DoubleArray) : DoubleArray
-    fun ToYCbCr601(values: DoubleArray) : DoubleArray
-    fun ToYCbCr709(values: DoubleArray) : DoubleArray
-    fun ToYCoCg(values: DoubleArray) : DoubleArray
+    fun ToRGB(values : FloatArray) : FloatArray
+    fun ToCMY(values : FloatArray) : FloatArray
+    fun ToHSL(values : FloatArray) : FloatArray
+    fun ToHSV(values: FloatArray) : FloatArray
+    fun ToYCbCr601(values: FloatArray) : FloatArray
+    fun ToYCbCr709(values: FloatArray) : FloatArray
+    fun ToYCoCg(values: FloatArray) : FloatArray
 }
